server/requirements.txt
# server/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
aiomysql==0.2.0
pydantic==2.4.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.0.0
cryptography==41.0.4
bcrypt==4.0.1
python-dotenv==1.0.0
sqlalchemy==2.0.23
sqlalchemy[asyncio]==2.0.23
# greenlet==2.0.2
alembic==1.12.1
pymysql